[refactor/postsNetworkBoundResource] refactor: update Post and related components to use Long for IDs for better compatibility

+++ feature/posts/src/main/java/com/alxnophis/jetpack/posts/data/datasource/PostsDataSource.kt
@@ -8,5 +8,5 @@ import com.alxnophis.jetpack.posts.data.model.PostsError
 internal interface PostsDataSource {
     suspend fun getPosts(): Either<PostsError, List<Post>>
 
-    suspend fun getPostById(postId: Int): Either<PostDetailError, Post>
+    suspend fun getPostById(postId: Long): Either<PostDetailError, Post>
 }
